New Features && Bug Fixes
List of Changes:
 - If command is on cooldown the remaining time will be sent to the user as a reply.
 - Added cooldown per user feature for chat commands.
 - Since now, no-command-identifier commands take the entire message as their command text. (TwitchAPI)

+++ src/bot/shared/Declensioner.cs
@@ -56,7 +56,7 @@ public static class Declensioner {
                };
     }
     
-    public static string Secs(int secs) {
+    public static string Secs(long secs) {
         var lastTwoDigits = secs%100;
         var lastDigit = secs%10;
